refactor(dashboard): replace AdminService with DashboardService
- migrated dashboard logic to a dedicated DashboardService for better separation of concerns
- moved DTOs to standalone files under `dto` folder for improved maintainability
- updated DashComponent to use DashboardService instead of AdminService
- removed AdminService and its associated test file as they are no longer needed
