+++ src/app/dto/PageDto.ts
@@ -1,0 +1,11 @@
+export interface PageDto<T> {
+  content: T[];
+  page: number;
+  size: number;
+  totalPages: number;
+  totalElements: number;
+  hasPrevious: boolean;
+  hasNext: boolean;
+  isFirst: boolean;
+  isLast: boolean;
+}
