+++ src/app/services/dashboard.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { AdminService } from './admin.service';
+import { DashboardService } from './dashboard.service';
 
-describe('AdminService', () => {
-  let service: AdminService;
+describe('DashboardService', () => {
+  let service: DashboardService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(AdminService);
+    service = TestBed.inject(DashboardService);
   });
 
   it('should be created', () => {
