+++ src/app/services/dashboard.service.ts
@@ -1,46 +1,15 @@
 import {Injectable} from '@angular/core';
-import {HttpClient, HttpParams} from '@angular/common/http';
-import {Observable} from 'rxjs';
+import {HttpClient, HttpParams} from "@angular/common/http";
 import {Globals} from "../globals/globals";
+import {Observable} from "rxjs";
+import {AffiliateStatsDto, DashboardDto, PaymentDataDto} from "../dto/DashboardDtos";
+import {PageDto} from "../dto/PageDto";
 
-export interface DashboardDto {
-  totalrevenue: number;
-  usercount: number;
-  allowlistcount: number;
-}
+@Injectable({
+  providedIn: 'root'
+})
+export class DashboardService {
 
-export interface PageDto<T> {
-  content: T[];
-  page: number;
-  size: number;
-  totalPages: number;
-  totalElements: number;
-  hasPrevious: boolean;
-  hasNext: boolean;
-  isFirst: boolean;
-  isLast: boolean;
-}
-
-export interface PaymentDataDto {
-  id: number;
-  paymentId: string;
-  payerId: number;
-  paymentStatus: string;
-  amount: number;
-  affiliateCode: string;
-}
-
-export interface AffiliateStatsDto {
-  firstname: string;
-  lastname: string;
-  affiliatecode: string;
-  roles: string[];
-  totalRevenue: number;
-  totalrewards: number;
-}
-
-@Injectable({providedIn: 'root'})
-export class AdminService {
   private baseUrl = this.globals.backendUrl + '/admin';
   private defaultPageSize = 10;
 
@@ -56,13 +25,14 @@ export class AdminService {
     if (status) {
       params = params.set('status', status);
     }
-    return this.http.get<PageDto<PaymentDataDto>>(`${this.baseUrl}/paymentdata`, { params });
+    return this.http.get<PageDto<PaymentDataDto>>(`${this.baseUrl}/paymentdata`, {params});
   }
 
   getAffiliateData(page = 0, size = this.defaultPageSize): Observable<PageDto<AffiliateStatsDto>> {
     const params = new HttpParams().set('page', page).set('size', size);
     return this.http.get<PageDto<AffiliateStatsDto>>(`${this.baseUrl}/affiliate`, {params});
   }
+
   get defaulPageSize() {
     return this.defaultPageSize;
   }
